♻️ Move logger.got from opts to internal/wrap
