+++ internal/wrap/logger.go
@@ -1,17 +1,19 @@
-package opts
+package wrap
 
 import (
 	"context"
 	"log/slog"
 	"time"
+
+	"github.com/akm/sql-slog/opts"
 )
 
 type SQLLogger struct {
 	*slog.Logger
-	Options *Options
+	Options *opts.Options
 }
 
-func NewSQLLogger(rawLogger *slog.Logger, opts *Options) *SQLLogger {
+func NewSQLLogger(rawLogger *slog.Logger, opts *opts.Options) *SQLLogger {
 	return &SQLLogger{
 		Logger:  rawLogger,
 		Options: opts,
@@ -22,11 +24,11 @@ func (x *SQLLogger) With(kv ...interface{}) *SQLLogger {
 	return NewSQLLogger(x.Logger.With(kv...), x.Options)
 }
 
-func (x *SQLLogger) StepWithoutContext(step *StepOptions, fn func() (*slog.Attr, error)) (*slog.Attr, error) {
+func (x *SQLLogger) StepWithoutContext(step *opts.StepOptions, fn func() (*slog.Attr, error)) (*slog.Attr, error) {
 	return x.Step(context.Background(), step, fn)
 }
 
-func (x *SQLLogger) Step(ctx context.Context, step *StepOptions, fn func() (*slog.Attr, error)) (*slog.Attr, error) {
+func (x *SQLLogger) Step(ctx context.Context, step *opts.StepOptions, fn func() (*slog.Attr, error)) (*slog.Attr, error) {
 	x.Log(ctx, slog.Level(step.Start.Level), step.Start.Msg)
 	t0 := time.Now()
 	attr, err := fn()
@@ -59,15 +61,15 @@ func (x *SQLLogger) Step(ctx context.Context, step *StepOptions, fn func() (*slo
 func (x *SQLLogger) durationAttr(d time.Duration) slog.Attr {
 	key := x.Options.DurationKey
 	switch x.Options.DurationType {
-	case DurationNanoSeconds:
+	case opts.DurationNanoSeconds:
 		return slog.Int64(key, d.Nanoseconds())
-	case DurationMicroSeconds:
+	case opts.DurationMicroSeconds:
 		return slog.Int64(key, d.Microseconds())
-	case DurationMilliSeconds:
+	case opts.DurationMilliSeconds:
 		return slog.Int64(key, d.Milliseconds())
-	case DurationGoDuration:
+	case opts.DurationGoDuration:
 		return slog.Duration(key, d)
-	case DurationString:
+	case opts.DurationString:
 		return slog.String(key, d.String())
 	default:
 		return slog.Int64(key, d.Nanoseconds())
