--- internal/wrap/sql_logger_aliases.go
@@ -1,14 +1,0 @@
-package wrap
-
-import sqlslogopts "github.com/akm/sql-slog/opts"
-
-type (
-	SQLLogger = sqlslogopts.SQLLogger
-)
-
-var (
-	NewSQLLogger = sqlslogopts.NewSQLLogger
-
-	IgnoreAttr  = sqlslogopts.IgnoreAttr
-	WithNilAttr = sqlslogopts.WithNilAttr
-)
